combine(update): merge into singular update script :recycle:
