--- scripts/lang/__init__.py
@@ -1,6 +1,0 @@
-from lang.dart import Dart
-from lang.go import Go
-from lang.java import Java
-from lang.python import Python
-
-__all__ = ["Dart", "Go", "Java", "Python"]
