--- scripts/lang/go.py
@@ -1,11 +1,0 @@
-from lang.base import LanguageBase
-
-
-class Go(LanguageBase):
-    """
-    Go implementation of LanguageBase.
-    """
-
-    def update_frugal(self, version):
-        # No change necessary
-        pass
